+++ identity/client/src/pages/users/modals/EditModal.tsx
@@ -14,7 +14,7 @@ import useTranslate from "src/utility/localization";
 import Divider from "src/components/form/Divider";
 import { FormModal, UseEntityModalProps } from "src/components/modal";
 import { updateUser, User } from "src/utility/api/users";
-import { isValidEmail } from "src/utility/isValidEmail";
+import { isValidEmail } from "src/utility/validate";
 
 type FormData = Pick<User, "name" | "email"> & {
   password: string;
